chore(instance): consolidate migrations, drop dead config key, relocate settings
- Merge instance_role column migration into create_instance_settings_table
  (shared timestamp, same feature); down() reverses both in order.
- Remove dead kit.workspaces.can_create_workspaces config key — the gate now
  flows through InstanceSettings. WORKSPACES_CAN_CREATE_WORKSPACE env is still
  honored as the seed default for workspace_creation_enabled via config/instance.php.
- Move InstanceSettings out of new app/Settings/ base folder into existing
  app/Support/; update namespace and all imports.
