+++ app/Providers/FortifyServiceProvider.php
@@ -6,7 +6,7 @@ use App\Actions\Fortify\CreateNewUser;
 use App\Actions\Fortify\ResetUserPassword;
 use App\Enums\SocialProvider;
 use App\Models\WorkspaceInvitation;
-use App\Settings\InstanceSettings;
+use App\Support\InstanceSettings;
 use Illuminate\Cache\RateLimiting\Limit;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\RateLimiter;
